requirements.txt
streamlit
pillow
fpdf